Extract type
